app/src/pyproject.toml
[tool.poetry]
name =  "Elyte FastAPI Application."
version = "0.1.0"
description = "e-Market"
authors = ["elyte5star <checkuti@gmail.com>"]
license = "Proprietary"
readme = "README.md"
packages = [{include = "fastapi_docker"}]
homepage ="https://github.com/elyte5star/fastapi_docker"
repository = "https://github.com/elyte5star/fastapi_docker"



classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Proprietary",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[elyte.api.app]
name =  "Elyte FastAPI Application."
version = "0.1.0"
description = """
   #   Interactive Documentation for Elyte e-Market
   """

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
pydantic = "^1.9.0"
uvicorn = {extras = ["standard"], version = "^0.17.1"}
pyconfs = "^0.5.5"
toml = "^0.10.2"
aiofiles = "^0.8.0"
motor = "^2.5.1"
asyncpg = "^0.25.0"
zipstream = "^1.1.4"
bcrypt = "^3.2.0"
jwt = "^1.3.1"
email-validator = "^1.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
aio-pika = "^7.1.2"
requests="^2.24.0"
itsdangerous="^2.1.2"
authlib="^1.0.1"
httpx="^0.23.0"
Jinja2="3.1.2"
sqlalchemy="1.4.44"
asyncmy="0.2.5"
newsapi-python="latest"
python-dateutil="2.8.2"
[tool.poetry.dev-dependencies]
black = "21.9b0"
isort = "^5.10.1"
mypy = "^0.931"
pydocstyle = "^6.1.1"


[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
[tool.isort]
profile = "black"
known_third_party = ["fastapi", "pydantic", "starlette"]
line_length = 120
float_to_top = true
lines_between_types = 0
lines_between_sections = 1
lines_after_imports = 2
force_single_line = false
use_parentheses = true
include_trailing_comma = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.mypy]
plugins = [
  "pydantic.mypy"
]

[database]
host = "db"
port = 3306
user = "userExample"
pwd = "54321"
db = "elyte"

[queue.params]
host_name = "rabbitmQ"
port = "5672"
my_queue ="db_operation"
pwd="elyteRQ"

[api]
log_type = "INFO"
host_url = "http://localhost:8080"
debug = true
auth_type= ["SELF","GOOGLE"]
google_client_id="864725361364-enrs63jgolid0jpmiffcsnj1rtadtss1.apps.googleusercontent.com"
news_api_key="8820ff85814d4e2ab94b0585d76eb209"

[admin]
email="checkuti@gmail.com"
username="elyte"
password="$2b$10$rQcvrrW2JcvjV2XM5TG3zeJd6oHPthld3VfRLsvyV2UJFO0.BxACO"
is_admin=true
telephone=40978057





[hash.password]
length = 16
rounds = 10
coding = "utf-8"

[token.params]
algorithm = "HS256"
secret_key = "0c2214c33cc65769166ec5248bb0ec6a15e892ba649e36fefc5732d9c1ba100"
token_url = "/token"
token_expire_min = 30
refresh_token_expire_minutes=43200 #30 days
grant_type="refresh_token"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
